Refactor

+++ Astrarium.Plugins.MinorBodies/AsteroidsCalc.cs
@@ -104,10 +104,10 @@ namespace Astrarium.Plugins.MinorBodies
             return MinorBodyEphem.Magnitude(a.G, a.H, beta, r, delta);
         }
 
-        private double Semidiameter(SkyContext c, Asteroid a)
+        private float Semidiameter(SkyContext c, Asteroid a)
         {
             double delta = c.Get(DistanceFromEarth, a);
-            return MinorBodyEphem.Semidiameter(delta, a.PhysicalDiameter);
+            return (float)MinorBodyEphem.Semidiameter(delta, a.PhysicalDiameter);
         }
 
         public void ConfigureEphemeris(EphemerisConfig<Asteroid> e)
